Island: Move propagation_credentials.py to credentials dir
