+++ monkey/monkey_island/cc/resources/credentials/propagation_credentials.py
